backend/requirements.txt
fastapi
uvicorn
paho-mqtt
httpx
sse-starlette
sqlalchemy
python-dotenv
pydantic
pydantic-settings
asyncio
